Potentially Fixed bug where Promises would overlap on their actions

+++ src/app/services/game/gameloop/game-loop.service.ts
@@ -9,8 +9,8 @@ import { GameAction } from 'src/app/models/game/GameAction';
 import { Unit } from 'src/app/models/game/units/Unit';
 import { BlockService } from '../../program-construction/block.service';
 import { LevelDataInterfaceService } from '../levelDataInterface/level-data-interface.service';
-import {Wait} from '../../../models/blockCommands/blocks/executable/Wait';
-import {UnitReadOnly} from '../../../models/game/units/UnitReadOnly';
+import { Wait } from '../../../models/blockCommands/blocks/executable/Wait';
+import { UnitReadOnly } from '../../../models/game/units/UnitReadOnly';
 
 @Injectable({
   providedIn: 'root'
@@ -124,6 +124,11 @@ export class GameLoopServiceService {
         this.grid[u.location.x][u.location.y] = u;
       }
 
+      console.log("Team 1 Data");
+      console.log(JSON.stringify(this.team1units));
+      console.log(JSON.stringify(this.team2units));
+      console.log(JSON.stringify(this.grid));
+
     } catch (error) {
       console.log("Failed: " + error);
       return false;
@@ -137,15 +142,59 @@ export class GameLoopServiceService {
    */
   stepGame(): Promise<any> {
     var self = this;
-    return Promise.race([this.baseStepPromise(), new Promise<GameAction>((resolve, reject) => {
-      setTimeout(function(){
-        resolve(new GameAction("RanOutOfTimeError", ((self.isTeam1Active) ? self.team1units : self.team2units)[self.unitIndex], null, false));
+    // return Promise.race([this.baseStepPromise(), new Promise<GameAction>((resolve, reject) => {
+    //   setTimeout(function () {
 
-        if(self.workerRunning != null) {
-          self.workerRunning.terminate();
+    //     var curTeam: Unit[] = (self.isTeam1Active) ? self.team1units : self.team2units;
+
+    //     self.unitIndex++;
+    //     self.codeIndex = 0;
+    //     if (curTeam.length <= self.unitIndex) {
+    //       //no more units to run through switch sides
+    //       self.unitIndex = 0;
+    //       self.isTeam1Active = !self.isTeam1Active;
+    //     }
+
+    //     self.currentConditions.clear();
+
+    //     if (self.workerRunning != null) {
+    //       self.workerRunning.terminate();
+    //     }
+
+    //     resolve(new GameAction("RanOutOfTimeError", ((self.isTeam1Active) ? self.team1units : self.team2units)[self.unitIndex], null, false));
+
+
+    //   }, this.WORKER_TIMEOUT_TIME);
+    // })]);
+    return new Promise<any>((resolve, reject) => {
+
+      Promise.race([this.baseStepPromise(), new Promise<GameAction>((resolve, reject) => {
+        setTimeout(function() {
+          resolve(new GameAction("RanOutOfTimeError", ((self.isTeam1Active) ? self.team1units : self.team2units)[self.unitIndex], null, false));
+        }, this.WORKER_TIMEOUT_TIME);
+      })]).then(result => {
+        var resolvedAction = result as GameAction;
+
+        if (resolvedAction.actionId === "RanOutOfTimeError") {
+          var curTeam: Unit[] = (self.isTeam1Active) ? self.team1units : self.team2units;
+
+          self.unitIndex++;
+          self.codeIndex = 0;
+          if (curTeam.length <= self.unitIndex) {
+            //no more units to run through switch sides
+            self.unitIndex = 0;
+            self.isTeam1Active = !self.isTeam1Active;
+          }
+
+          self.currentConditions.clear();
+
+          if (self.workerRunning != null) {
+            self.workerRunning.terminate();
+          }
         }
-      }, this.WORKER_TIMEOUT_TIME);
-    })]);
+        resolve(result);
+      });
+    });
   }
 
   baseStepPromise(): Promise<GameAction> {
@@ -186,21 +235,22 @@ export class GameLoopServiceService {
 
           this.lastAction = null;
 
-           //check for death
-           if(last.hasDied) {
-            var team: Unit[] = (last.receiver.team == 1) ? this.team1units : this.team2units;
-            var indexDead = team.indexOf(last.receiver);
-            if(indexDead >= 0 && indexDead < team.length) {
-              team.splice(indexDead, 1);
+          //check for death
+          if (last.hasDied) {
+            // var team: Unit[] = (last.receiver.team == 1) ? this.team1units : this.team2units;
+            // var indexDead = team.indexOf(last.receiver);
+            // if(indexDead >= 0 && indexDead < team.length) {
+            //   team.splice(indexDead, 1);
 
-              //adjust new code index
-              if(this.unitIndex > indexDead) {
-                this.codeIndex--;
-              }
+            //   //adjust new code index
+            //   if(this.unitIndex > indexDead) {
+            //     this.codeIndex--;
+            //   }
 
-            } else {
-              console.log("You shouldnt be seeing this error message");
-            }
+            // } else {
+            //   console.log("You shouldnt be seeing this error message");
+            // }
+            this.deleteUnit(last.receiver);
           }
 
           //next unit
@@ -217,16 +267,16 @@ export class GameLoopServiceService {
           this.currentConditions.clear();
 
           successFunc(last);
-        } else if(unit.codeType === CodeType.FILE) {
+        } else if (unit.codeType === CodeType.FILE) {
 
           this.workerRunning = unit.activecode as Worker;
 
-          this.workerRunning.postMessage(JSON.stringify({grid: this.convertGridToReadOnly(this.grid), unit: new UnitReadOnly(unit)}));
+          this.workerRunning.postMessage(JSON.stringify({ grid: this.convertGridToReadOnly(this.grid), unit: new UnitReadOnly(unit) }));
 
           var self = this
           var messageSent = false;
 
-          this.workerRunning.onmessage = function(event) {
+          this.workerRunning.onmessage = function (event) {
 
             if (!messageSent) {
               self.workerRunning = null;
@@ -240,19 +290,20 @@ export class GameLoopServiceService {
 
               //check for death
               if (last.hasDied) {
-                var team: Unit[] = (last.receiver.team == 1) ? self.team1units : self.team2units;
-                var indexDead = team.indexOf(last.receiver);
-                if (indexDead >= 0 && indexDead < team.length) {
-                  team.splice(indexDead, 1);
+                // var team: Unit[] = (last.receiver.team == 1) ? self.team1units : self.team2units;
+                // var indexDead = team.indexOf(last.receiver);
+                // if (indexDead >= 0 && indexDead < team.length) {
+                //   team.splice(indexDead, 1);
 
-                  //adjust new code index
-                  if (self.unitIndex > indexDead) {
-                    self.codeIndex--;
-                  }
+                //   //adjust new code index
+                //   if (self.unitIndex > indexDead) {
+                //     self.codeIndex--;
+                //   }
 
-                } else {
-                  console.log("You shouldnt be seeing this error message");
-                }
+                // } else {
+                //   console.log("You shouldnt be seeing this error message");
+                // }
+                self.deleteUnit(last.receiver);
               }
 
               var curTeam: Unit[] = (self.isTeam1Active) ? self.team1units : self.team2units;
@@ -270,7 +321,7 @@ export class GameLoopServiceService {
             messageSent = true;
           }
 
-          this.workerRunning.onerror = function(event) {
+          this.workerRunning.onerror = function (event) {
             self.workerRunning = null;
 
             var curTeam: Unit[] = (self.isTeam1Active) ? self.team1units : self.team2units;
@@ -382,14 +433,14 @@ export class GameLoopServiceService {
 
   }
 
-  private convertGridToReadOnly(grid: Unit[][]){
+  private convertGridToReadOnly(grid: Unit[][]) {
 
     let newGrid = [];
-    for (let row of grid){
+    for (let row of grid) {
       let newRow = [];
-      for(let el of row){
+      for (let el of row) {
 
-        if(el === null){
+        if (el === null) {
           newRow.push(el);
           continue;
         }
@@ -463,37 +514,38 @@ export class GameLoopServiceService {
    * @param unit the unit the user is controlling
    * @private
    */
-  private convertWorkerMessageToAction(data, grid, unit): GameAction{
+  private convertWorkerMessageToAction(data, grid, unit): GameAction {
 
     let action = data.result;
     try {
       let executable = this.blockServ.getById(btoa(action));
-      if(!(this.blockServ.isExecutable(executable))){
+      if (!(this.blockServ.isExecutable(executable))) {
         throw new Error();
-      }else{
+      } else {
         return executable.execute(grid, unit);
       }
 
-    }catch (err){
+    } catch (err) {
 
       return new Wait().execute(grid, unit);
 
     }
   }
 
-  deleteUnit(unit: Unit){
+  deleteUnit(unit: Unit) {
 
-    let team = unit.team === 1 ? this.team1units : this.team2units;
-    for(let i = 0; i < team.length; i++){
-
-      if(team[i].id === unit.id){
-
+    console.log("Deleting unit " + unit.id + " of team " + unit.team);
+    var team = (unit.team === 1 ? this.team1units : this.team2units);
+    console.log("Looked at team " + team + " of type " + JSON.stringify((team === this.team1units ? "1" : "2")));
+    for (let i = 0; i < team.length; i++) {
+      if (team[i].id === unit.id) {
+        console.log("Deleted unit " + unit.id + " at index " + i);
         team.splice(i);
-
       }
-
     }
 
+    this.grid[unit.location.x][unit.location.y] = null;
+    console.log("end");
   }
 
 }
